fill in the "where" action.
Make it so all the tests show up on the index page.
stupid to keep changing the test in the controller.

Filled in the "where" action.
It does okay, and is mostly readable until  you try to start sorting.

You can only sort on properties of kitten.
Not sure what the special technique is for sorting on other properties.